+++ test/foundry/Token.t.sol
@@ -408,7 +408,7 @@ contract TokenTest is Test {
     }
 
     function test_version() public view {
-        assertEq(token.version(), "2.4.0");
+        assertEq(token.version(), "2.5.0");
     }
 
     /**